Integrated sinle Transaction Service and Transaction Controller class

+++ src/main/java/com/example/walletapplication/repository/WalletRepository.java
@@ -1,8 +1,9 @@
 package com.example.walletapplication.repository;
 
-import com.example.walletapplication.entity.User;
 import com.example.walletapplication.entity.Wallet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletRepository extends JpaRepository<Wallet, Integer> {
+
+public interface WalletRepository extends JpaRepository<Wallet, Long> {
+
 }
